% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Caoband.R
\name{Caoband}
\alias{Caoband}
\title{Cao and Cao2 confidence bands for functional means}
\usage{
Caoband(
  Y,
  as,
  checkpts = rep(1:100)/100 * ub,
  ub = max(as),
  mean_true,
  c = 0.5,
  p = 2,
  b_M = 1000,
  alpha = c(0.05, 0.01),
  N_m = NA,
  cov_nnd = 1
)
}
\arguments{
\item{Y}{a matrix with the i-th row representing the i-th subject's observed functional data (e.g., occupation time data), the j-th column corresponding to the observed functional data at the j-th design time point (e.g., activity level) for each subject}

\item{as}{a vector consisting of the design time points of the observed functional data; that is, \eqn{{\bf G}_{n}} (from the smallest to the largest).}

\item{checkpts}{a vector consisting of the time points (from the smallest to the largest; e.g., the activity levels of interest) at which we want to evaluate the confidence band.}

\item{ub}{the largest time point (e.g., the activity levels of interest) at which we want to evaluate the confidence band.}

\item{mean_true}{a vector containing the true functional mean over \code{checkpts} (defined above)}

\item{c}{a parameter for controlling the number of interior knots in estimating the functional mean, as described in Section 4 of Cao et al. (2012) (default = 0.5)}

\item{p}{the order of spline functions (default = 2)}

\item{b_M}{the number of copies in simulating \eqn{\zeta(x)} for constructing the confidence bounds, as described in Section 4 of Cao et al. (2012) (default = 0.5) (default = 1000)}

\item{alpha}{a vector of significance levels of interest (default = c(0.05, 0.01))}

\item{N_m}{the number of interior knots in estimating the functional mean, as described in Section 4 of Cao et al. (2012)}

\item{cov_nnd}{1 if the initially smoothed covariance estimates are projected onto the space of non-negative definite matrices and 0 if not; -1 if not doing the aforementioned projection and not taken into account the fact that the resulting estimated covariance can have complex eigenvalues}
}
\value{
a list containing the following:
\item{cover_or_not}{a \code{length(alpha)}-vector with the \code{i}-th element indicating if the true functional mean is covered by the \eqn{100(1-}\code{alpha[i]}\eqn{)\%} confidence band (1 if not covered and 0 otherwise)}
\item{mean}{a vector containing the estimated mean function using spline estimation}
\item{confidence_bounds}{a matrix containing the confidence bounds. The (\code{2 * i - 1})-th and (\code{2 * i})-th columns contain the upper and lower bound of the \eqn{100(1-}\code{alpha[i]}\eqn{)\%} confidence band, respectively}
\item{covariance}{the estimated covariance of the functional data; the initially smoothed covariance estimates are projected onto the space of non-negative definite matrices}
\item{number_of_noises}{the number of eigenfunctions described in Section 4 of Cao et al. (2012)}
\item{noises}{a matrix containing the \eqn{\hat{\phi}_{k}(\cdot)} functions defined in Section 4 of Cao et al. (2012), with the \eqn{k}-th column containing \eqn{\hat{\phi}_{k}(\cdot)}}
\item{confidence_level}{\code{1 - alpha}}
\item{runtime}{the runtime of the \code{Caoband} funciton}
}
\description{
Compute the Cao and Cao2 confidence bands modified from the bands by Cao et al. (2012), as mentioned in Section 3, along with whether they capture the true functional mean or not.
}
\examples{
CaoCB <- Caoband(Y = OTdata$Ta[[1]], as = OTdata$as, checkpts = OTdata$as, mean_true = rep(0, 500))
CaoCB$confidence_bounds
}
\references{
Cao, G., Yang, L. and Todem, D. (2012) Simultaneous inference for the mean function based on dense functional data. \emph{Journal of Nonparametric Statistics}, \strong{24}, 359--377.
}
\author{
The code is first written by Yan-Yu Chen on 2021/02/03 and later modified slightly by Hsin-wen Chang.
}
