% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elfanova.R
\name{elfanova}
\alias{elfanova}
\title{EL and related functional ANOVA tests}
\usage{
elfanova(Ta, as, n_boot = 1000, as_right_ind_keep_as0 = length(as))
}
\arguments{
\item{Ta}{a list with the j-th element being the matrix of observed functional data (e.g., occuptation time curve) for the j-th group, and the i-th row of that matrix representing the observed curve of the i-th subject in the j-th group}

\item{as}{a vector consisting of the design time points (e.g., activity levels) of the observed functional data; that is, \eqn{{\bf G}_n} defined in Section 2.2 (from the smallest to the largest).}

\item{n_boot}{number of bootstrap samples.}

\item{as_right_ind_keep_as0}{the minimal (among the \code{length(Ta)} groups) index in the vector \code{as} (defined above) corresponding to each element of the set \eqn{M} (defined in Suppplement Section 5.2).}
}
\value{
a list containing the following:
\item{suptest}{test statistic of the proposed maximally selected EL test}
\item{inttest_da}{test statistic of the integrated EL test}
\item{suptest_EP}{test statistic of the Wald-type test that is asymptotically equivalent to the proposed maximally selected EL test}
\item{inttest_EP_da}{test statistic of the Wald-type test that is asymptotically equivalent to the integrated EL test}
\item{out_sup_pval}{p-value of the proposed maximally selected EL test}
\item{out_dF_pval}{p-value of the integrated EL test}
\item{out_sup_EP_pval}{p-value of the Wald-type test that is asymptotically equivalent to the proposed maximally selected EL test}
\item{out_dF_EP_pval}{p-value of the Wald-type test that is asymptotically equivalent to the integrated EL test}
\item{as}{the input \code{as}; see Arguments}
\item{mu_hat_vec}{a \code{length(Ta)} x \code{length(as)} matrix; j-th row is the sample mean of \code{Ta[[j]]}}
\item{supboot}{bootstrap values of the proposed maximally selected EL test}
\item{int_da_boot}{bootstrap values of the integrated EL test}
}
\description{
Computes EL-based functional ANOVA tests and the corresponding asymptotically equivalent Wald-type tests
}
\examples{
set.seed(1008, kind = "Mersenne-Twister", normal.kind = "Inversion", sample.kind = "Rejection")
test_out = elfanova(Ta = OTdata$Ta, as = OTdata$as)
test_out$out_sup_pval
}
