% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mu_a_hat.R
\name{mu_a_hat}
\alias{mu_a_hat}
\title{Occupation time curve}
\usage{
mu_a_hat(Xt, grid_widths, as)
}
\arguments{
\item{Xt}{a vector of raw activity data on a grid of time points}

\item{grid_widths}{a vector showing the number of time units (can be \eqn{< 1}) each grid point contains. This vector should have the same length as \code{Xt}.}

\item{as}{a vector showing the grid of activity levels over which the occupation time is evaluated.}
}
\value{
a vector containing the observed occupation time curve based on \code{Xt}
}
\description{
Computes the occupation time curve based on the raw activity data
}
