% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rawlens.R
\docType{data}
\name{rawlens}
\alias{rawlens}
\title{Lengths of raw activity data}
\format{
A list of four elements corresponding to the four groups: veterans aged 75-and-older, non-veterans aged 75-and-older, veterans aged 65--74, and non-veterans aged aged 65--74. Each element contains a vector recording the length of non-NA raw activity data from each subject.
}
\source{
\href{https://wwwn.cdc.gov/nchs/nhanes/continuousnhanes/default.aspx?BeginYear=2005}{National Health and Nutrition Examination Survey Data, 2005--2006}
}
\usage{
rawlens
}
\description{
Lengths of non-NA raw activity data from subjects aged 65-and-older in the 2005--06 U.S. National Health and Nutrition Examination Survey
}
\keyword{datasets}
